backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic-settings==2.1.0

# ── Database ───────────────────────────────────────────────────────────────
sqlalchemy==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# ── Auth ───────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ── LLM / API clients ──────────────────────────────────────────────────────
openai==1.9.0
httpx==0.26.0

# ── Document parsing ───────────────────────────────────────────────────────
PyPDF2==3.0.1
pdfplumber==0.11.0          # 高精度 PDF 表格提取，支持多列布局
python-docx==1.1.0
openpyxl==3.1.2
xlrd==2.0.1
xlsxwriter==3.2.0           # Excel 生成与格式化输出
python-pptx==0.6.23
lxml==5.1.0
chardet==5.2.0
python-magic==0.4.27

# ── Data science core ──────────────────────────────────────────────────────
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4               # 统计检验、时间序列分析、信号处理
scikit-learn==1.3.2         # 聚类、异常检测、特征重要性
statsmodels==0.14.1         # OLS 回归、ARIMA 时序、假设检验

# ── Data presentation ──────────────────────────────────────────────────────
tabulate==0.9.0             # 控制台 / Markdown 表格格式化
jinja2==3.1.4               # 报告模板引擎

# ── Visualization ──────────────────────────────────────────────────────────
matplotlib==3.8.2
seaborn==0.13.1
Pillow==10.2.0

# ── Security / Sandbox ─────────────────────────────────────────────────────
RestrictedPython==7.0

# ── Async / IO ─────────────────────────────────────────────────────────────
aiofiles==23.2.1
websockets==12.0

# ── Utilities ──────────────────────────────────────────────────────────────
redis==5.0.1
